Massively reorganized alchemy. added basic unit tests
