--- Alchemy.config
@@ -1,82 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
--->
-
-<configuration>
-  <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler,
-            log4net-net-1.0" />
-  </configSections>
-
-  <log4net>
-
-    <logger name="Alchemy.Log">
-      <level value="INFO" />
-    </logger>
-
-    <appender name="LogFileAppender" type="log4net.Appender.RollingFileAppender">
-      <param name="File" value="Alchemy.Log" />
-      <param name="AppendToFile" value="true" />
-      <rollingStyle value="Size" />
-      <maxSizeRollBackups value="10" />
-      <maximumFileSize value="10MB" />
-      <staticLogFileName value="true" />
-      <layout type="log4net.Layout.PatternLayout">
-        <param name="ConversionPattern" value="%-5p%d{yyyy-MM-dd hh:mm:ss} – %m%n" />
-      </layout>
-    </appender>
-
-    <appender name="ConsoleAppender" type="log4net.Appender.ColoredConsoleAppender">
-      <mapping>
-        <level value="INFO" />
-        <foreColor value="White, HighIntensity" />
-        <backColor value="Green" />
-      </mapping>
-      <mapping>
-        <level value="DEBUG" />
-        <foreColor value="White, HighIntensity" />
-        <backColor value="Blue" />
-      </mapping>
-      <mapping>
-        <level value="WARN" />
-        <foreColor value="Yellow, HighIntensity" />
-        <backColor value="Purple" />
-      </mapping>
-      <mapping>
-        <level value="ERROR" />
-        <foreColor value="Yellow, HighIntensity" />
-        <backColor value="Red" />
-      </mapping>
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%message%newline" />
-      </layout>
-    </appender>
-
-    <root>
-      <level value="ALL" />
-      <appender-ref ref="LogFileAppender" />
-      <appender-ref ref="ConsoleAppender" />
-    </root>
-
-  </log4net>
-</configuration>
