+++ Alchemy.sln
@@ -1,0 +1,85 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alchemy", "src\Alchemy\Alchemy.csproj", "{45486CDE-86A3-4769-952F-E0821BF79493}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Startup", "..\Startup\Startup.csproj", "{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{32470B4C-0665-44B9-9A55-AF9C671FDAA5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8C9C334C-588E-41A1-85FC-9DC035821D27}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration", "{24B00C6C-88BE-4F83-B2E1-8CC8F145BFB9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Unit", "Unit", "{A8BCA98B-12E7-4E13-99FE-09DDCA10BB26}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Alchemy", "Alchemy", "{DEF8E4AE-D1A2-4C33-A4EC-A51BDF6BD09F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alchemy", "test\Unit\Alchemy\Alchemy.csproj", "{D62E2F92-A930-44D9-90FB-864CAF837BAE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|x64.ActiveCfg = Release|Any CPU
+		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|x86.ActiveCfg = Release|Any CPU
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x64.ActiveCfg = Debug|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x64.Build.0 = Debug|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x86.ActiveCfg = Debug|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x86.Build.0 = Debug|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Any CPU.ActiveCfg = Release|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Mixed Platforms.ActiveCfg = Release|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Mixed Platforms.Build.0 = Release|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x64.ActiveCfg = Release|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x64.Build.0 = Release|x64
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x86.ActiveCfg = Release|x86
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x86.Build.0 = Release|x86
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|x64.ActiveCfg = Release|Any CPU
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE}.Release|x86.ActiveCfg = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9} = {8C9C334C-588E-41A1-85FC-9DC035821D27}
+		{45486CDE-86A3-4769-952F-E0821BF79493} = {8C9C334C-588E-41A1-85FC-9DC035821D27}
+		{24B00C6C-88BE-4F83-B2E1-8CC8F145BFB9} = {32470B4C-0665-44B9-9A55-AF9C671FDAA5}
+		{A8BCA98B-12E7-4E13-99FE-09DDCA10BB26} = {32470B4C-0665-44B9-9A55-AF9C671FDAA5}
+		{DEF8E4AE-D1A2-4C33-A4EC-A51BDF6BD09F} = {A8BCA98B-12E7-4E13-99FE-09DDCA10BB26}
+		{D62E2F92-A930-44D9-90FB-864CAF837BAE} = {DEF8E4AE-D1A2-4C33-A4EC-A51BDF6BD09F}
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		StartupItem = Server.csproj
+	EndGlobalSection
+EndGlobal
