--- Server.csproj
@@ -1,80 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{45486CDE-86A3-4769-952F-E0821BF79493}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Alchemy.Server</RootNamespace>
-    <AssemblyName>Alchemy</AssemblyName>
-    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup>
-    <StartupObject />
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="log4net">
-      <HintPath>Libraries\log4net.dll</HintPath>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Web" />
-    <Reference Include="System.XML" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="AccessPolicyServer.cs" />
-    <Compile Include="Handlers\WebSocket\DataFrame.cs" />
-    <Compile Include="Handlers\WebSocket\hybi00\DataFrame.cs" />
-    <Compile Include="Handlers\WebSocket\hybi10\DataFrame.cs" />
-    <Compile Include="Classes\Header.cs" />
-    <Compile Include="Classes\Response.cs" />
-    <Compile Include="Classes\UserContext.cs" />
-    <Compile Include="Handlers\DefaultHandler.cs" />
-    <Compile Include="Handlers\Handler.cs" />
-    <Compile Include="Handlers\WebSocket\hybi00\WebSocketHandler.cs" />
-    <Compile Include="Handlers\WebSocket\hybi00\Handshakes.cs" />
-    <Compile Include="Handlers\WebSocket\hybi00\WebSocketAuthentication.cs" />
-    <Compile Include="Handlers\WebSocket\hybi10\WebSocketAuthentication.cs" />
-    <Compile Include="Handlers\WebSocket\hybi10\Handshakes.cs" />
-    <Compile Include="Handlers\WebSocket\hybi10\WebSocketHandler.cs" />
-    <Compile Include="Handlers\WebSocket\IWebSocketAuthentication.cs" />
-    <Compile Include="TcpServer.cs" />
-    <Compile Include="WebSocketServer.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Alchemy.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      <SubType>Designer</SubType>
-    </None>
-  </ItemGroup>
-  <ItemGroup />
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
