--- Server.sln
@@ -1,45 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Server", "Server.csproj", "{45486CDE-86A3-4769-952F-E0821BF79493}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Startup", "..\Startup\Startup.csproj", "{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Any CPU.Build.0 = Release|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{45486CDE-86A3-4769-952F-E0821BF79493}.Release|x86.ActiveCfg = Release|Any CPU
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x86.ActiveCfg = Debug|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Debug|x86.Build.0 = Debug|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Any CPU.ActiveCfg = Release|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Mixed Platforms.ActiveCfg = Release|x64
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|Mixed Platforms.Build.0 = Release|x64
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x86.ActiveCfg = Release|x86
-		{8F9A2800-57C7-4A96-84E3-24CE5AEC5EF9}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = Server.csproj
-	EndGlobalSection
-EndGlobal
