+++ src/Alchemy/Handlers/WebSocket/hybi10/Handshakes.cs
@@ -1,0 +1,160 @@
+﻿/*
+Copyright 2011 Olivine Labs, LLC.
+http://www.olivinelabs.com
+*/
+
+/*
+This file is part of Alchemy Websockets.
+
+Alchemy Websockets is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Alchemy Websockets is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using Alchemy.Classes;
+
+namespace Alchemy.Handlers.WebSocket.hybi10
+{
+    /// <summary>
+    /// An easy wrapper for the header to access client handshake data.
+    /// See http://www.whatwg.org/specs/web-socket-protocol/ for more details on the WebSocket Protocol.
+    /// </summary>
+    public class ClientHandshake
+    {
+        /// <summary>
+        /// The preformatted handshake as a string.
+        /// </summary>
+        private const String Handshake =
+            "GET {0} HTTP/1.1\r\n" +
+            "Host: {2}\r\n" +
+            "Upgrade: websocket\r\n" +
+            "Connection: Upgrade\r\n" +
+            "Sec-WebSocket-Key: {4}\r\n" +
+            "Sec-WebSocket-Origin: {1}\r\n" +
+            "Sec-WebSocket-Protocol: {3}\r\n" +
+            "Sec-WebSocket-Version: 8\r\n" +
+            "{5}";
+
+        public string Host = String.Empty;
+        public string Key = String.Empty;
+        public string Origin = String.Empty;
+        public string ResourcePath = String.Empty;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ClientHandshake"/> class.
+        /// </summary>
+        /// <param name="header">The header.</param>
+        public ClientHandshake(Header header)
+        {
+            ResourcePath = header.RequestPath;
+            Key = header["sec-websocket-key"];
+            SubProtocol = header["sec-websocket-protocol"];
+            Origin = header["origin"];
+            if (String.IsNullOrEmpty(Origin))
+            {
+                Origin = header["sec-websocket-origin"];
+            }
+            Host = header["host"];
+            Version = header["sec-websocket-version"];
+            Cookies = header.Cookies;
+        }
+
+        public HttpCookieCollection Cookies { get; set; }
+        public string SubProtocol { get; set; }
+        public string Version { get; set; }
+        public Dictionary<string, string> AdditionalFields { get; set; }
+
+        /// <summary>
+        /// Determines whether this instance is valid.
+        /// </summary>
+        /// <returns>
+        ///   <c>true</c> if this instance is valid; otherwise, <c>false</c>.
+        /// </returns>
+        public bool IsValid()
+        {
+            return (
+                       (Host != null) &&
+                       (Key != null) &&
+                       (Int32.Parse(Version) >= 8)
+                   );
+        }
+
+        /// <summary>
+        /// Returns a <see cref="System.String"/> that represents this instance.
+        /// </summary>
+        /// <returns>
+        /// A <see cref="System.String"/> that represents this instance.
+        /// </returns>
+        public override string ToString()
+        {
+            string additionalFields = String.Empty;
+
+            if (Cookies != null)
+            {
+                additionalFields += "Cookie: " + Cookies + "\r\n";
+            }
+
+            if (additionalFields != String.Empty)
+            {
+                additionalFields = AdditionalFields.Aggregate(additionalFields,
+                                                              (current, field) =>
+                                                              current + (field.Key + ": " + field.Value + "\r\n"));
+            }
+            additionalFields += "\r\n";
+
+            return String.Format(Handshake, ResourcePath, Origin, Host, SubProtocol, Key, additionalFields);
+        }
+    }
+
+    /// <summary>
+    /// Implements a server handshake
+    /// See http://www.whatwg.org/specs/web-socket-protocol/ for more details on the WebSocket Protocol.
+    /// </summary>
+    public class ServerHandshake
+    {
+        /// <summary>
+        /// The preformatted handshake string.
+        /// </summary>
+        private const string Handshake =
+            "HTTP/1.1 101 Switching Protocols\r\n" +
+            "Upgrade: websocket\r\n" +
+            "Connection: Upgrade\r\n" +
+            "Sec-WebSocket-Accept: {0}\r\n" +
+            "{1}";
+
+        public string Accept { get; set; }
+        public string SubProtocol { get; set; }
+
+        public Dictionary<string, string> AdditionalFields { get; set; }
+
+        /// <summary>
+        /// Returns a <see cref="System.String"/> that represents this instance.
+        /// </summary>
+        /// <returns>
+        /// A <see cref="System.String"/> that represents this instance.
+        /// </returns>
+        public override string ToString()
+        {
+            string additionalFields = String.Empty;
+            if (SubProtocol != null)
+            {
+                additionalFields += "Sec-WebSocket-Protocol: " + SubProtocol + "\r\n";
+            }
+            additionalFields += "\r\n";
+            return String.Format(Handshake, Accept, additionalFields);
+        }
+    }
+}
